webxr/anchors/ar_anchor_states.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/webxr_util.js"></script>
<script src="../resources/webxr_test_asserts.js"></script>
<script src="../resources/webxr_test_constants.js"></script>
<script src="../resources/webxr_test_constants_fake_world.js"></script>
<canvas />

<script>

// 1m above world origin.
const VIEWER_ORIGIN_TRANSFORM = {
  position: [0, 1, 0],
  orientation: [0, 0, 0, 1],
};

const fakeDeviceInitParams = {
  supportedModes: ["immersive-ar"],
  views: VALID_VIEWS,
  supportedFeatures: ALL_FEATURES,
  viewerOrigin: VIEWER_ORIGIN_TRANSFORM,
};

// Creates a test method that leverages anchors API.
// |shouldSucceed| - true if the anchors creation request is expected to succeed, false otherwise
// |endSession| - true if the test case should call session.end() prior to creating an anchor
// |expectedError| - expected error name that should be returned in case shouldSucceed is false
const testFunctionGenerator = function(shouldSucceed, endSession, expectedError) {

  const testFunction = function(session, fakeDeviceController, t) {

    let debug = xr_debug.bind(this, 'testAnchorStates');

    fakeDeviceController.addEventListener("anchorcreate", (anchorCreateEvent) => {
      anchorCreateEvent.success = true;
    });

    let watcherDone = new Event("watcherdone");
    let eventWatcher = new EventWatcher(t, session, ["watcherdone"]);
    let eventPromise = eventWatcher.wait_for(["watcherdone"]);

    session.requestReferenceSpace('local').then((localRefSpace) => {

      const onFrame = function(time, frame) {
        debug("rAF 1");

        if(endSession) {
          debug("ending session");
          session.end();
        }

        debug("creating anchor");
        frame.createAnchor(new XRRigidTransform(), localRefSpace)
          .then((anchor) => {
            debug("anchor created");

            t.step(() => {
              assert_true(shouldSucceed,
                "`createAnchor` succeeded when it was expected to fail");
            });

            session.dispatchEvent(watcherDone);
          }).catch((error) => {
            debug("anchor creation failed");

            t.step(() =>  {
              assert_false(shouldSucceed,
                "`createAnchor` failed when it was expected to succeed, error: " + error);
              assert_equals(error.name, expectedError,
                "`createAnchor` failed with unexpected error name");
            });

            session.dispatchEvent(watcherDone);
          });

        // Anchor result will only take effect with frame data - schedule
        // a frame after we requested anchor creation, otherwise the test will time out.
        session.requestAnimationFrame(() => {
          debug("rAF 2");
        });
      };

      debug("requesting animation frame");
      session.requestAnimationFrame(onFrame);
    }); // session.requestReferenceSpace(...)

    return eventPromise;
  };  // testFunction

  return testFunction;
};

xr_session_promise_test("Anchor creation succeeds if the feature was requested",
  testFunctionGenerator(/*shouldSucceed=*/true, /*endSession=*/false),
  fakeDeviceInitParams,
  'immersive-ar', { 'requiredFeatures': ['anchors'] });

xr_session_promise_test("Anchor creation fails if the feature was not requested",
  testFunctionGenerator(/*shouldSucceed=*/false, /*endSession=*/false, "NotSupportedError"),
  fakeDeviceInitParams,
  'immersive-ar', {});

xr_session_promise_test("Anchor creation fails if the feature was requested but the session already ended",
  testFunctionGenerator(/*shouldSucceed=*/false, /*endSession=*/true, "InvalidStateError"),
  fakeDeviceInitParams,
  'immersive-ar', { 'requiredFeatures': ['anchors'] });

</script>
