third_party/blink/web_tests/platform/win11-arm64/external/wpt/webxr/anchors/ar_anchor_states.https-expected.txt

This is a testharness.js-based test.
[FAIL] Anchor creation succeeds if the feature was requested - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Anchor creation succeeds if the feature was requested - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Anchor creation fails if the feature was not requested - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Anchor creation fails if the feature was not requested - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Anchor creation fails if the feature was requested but the session already ended - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Anchor creation fails if the feature was requested but the session already ended - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
